+++ crates/pytauri-core/src/lib.rs
@@ -19,9 +19,7 @@ pub mod ext_mod {
     use super::*;
 
     #[pymodule_export]
-    pub use ext_mod_impl::{
-        App, AppHandle, Context, Event, Listener, Manager, RunEvent, RunEventEnum,
-    };
+    pub use ext_mod_impl::{App, AppHandle, Context, Event, Listener, Manager, RunEvent};
 
     pub use ext_mod_impl::{
         EventId, ImplListener, ImplManager, PyAppHandleExt, PyAppHandleStateError,