+++ python/pytauri/src/pytauri/__init__.py
@@ -14,8 +14,7 @@ from pytauri.ffi import (
     Listener,
     Manager,
     RunEvent,
-    RunEventEnum,
-    RunEventEnumType,
+    RunEventType,
     builder_factory,
     context_factory,
 )
@@ -36,8 +35,7 @@ __all__ = [
     "Listener",
     "Manager",
     "RunEvent",
-    "RunEventEnum",
-    "RunEventEnumType",
+    "RunEventType",
     "builder_factory",
     "context_factory",
 ]