+++ crates/pyo3-utils/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Docs
+
+- [#57](https://github.com/WSH032/pytauri/pull/57) - add documentation to `mod py_match` indicating it should only be used for `state-machine-like` `enum`, not for `Union-like` `enum`.
+
 ## [0.1.0-beta.0]
 
 [unreleased]: https://github.com/WSH032/pytauri/tree/HEAD